+++ GRDB/Core/Support/Foundation/Date.swift
@@ -3,26 +3,75 @@ import Foundation
 /// NSDate is stored in the database using the format
 /// "yyyy-MM-dd HH:mm:ss.SSS", in the UTC time zone.
 extension NSDate : DatabaseValueConvertible {
-    
-    /// Returns a value that can be stored in the database.
+    /// Returns a database value that contains the date encoded as
+    /// "yyyy-MM-dd HH:mm:ss.SSS", in the UTC time zone.
     public var databaseValue: DatabaseValue {
-        return storageDateFormatter.string(from: self as Date).databaseValue
+        return (self as Date).databaseValue
     }
     
-    /// Returns a Date initialized from *dbValue*, if possible.
+    /// Returns a date initialized from dbValue, if possible.
+    ///
+    /// If database value contains a number, that number is interpreted as a
+    /// timeinterval since 00:00:00 UTC on 1 January 1970.
+    ///
+    /// If database value contains a string, that string is interpreted as a
+    /// [SQLite date](https://sqlite.org/lang_datefunc.html) in the UTC time
+    /// zone. Nil is returned if the date string does not contain at least the
+    /// year, month and day components. Other components (minutes, etc.)
+    /// are set to zero if missing.
     public static func fromDatabaseValue(_ dbValue: DatabaseValue) -> Self? {
-        if let databaseDateComponents = DatabaseDateComponents.fromDatabaseValue(dbValue) {
-            return cast(fromDatabaseDateComponents(databaseDateComponents))
+        guard let date = Date.fromDatabaseValue(dbValue) else {
+            return nil
         }
-        if let julianDayNumber = Double.fromDatabaseValue(dbValue) {
-            return cast(fromJulianDayNumber(julianDayNumber))
+        return cast(date)
+    }
+}
+
+/// Date is stored in the database using the format
+/// "yyyy-MM-dd HH:mm:ss.SSS", in the UTC time zone.
+extension Date : DatabaseValueConvertible {
+    /// Returns a database value that contains the date encoded as
+    /// "yyyy-MM-dd HH:mm:ss.SSS", in the UTC time zone.
+    public var databaseValue: DatabaseValue {
+        return storageDateFormatter.string(from: self).databaseValue
+    }
+    
+    /// Returns a date initialized from dbValue, if possible.
+    ///
+    /// If database value contains a number, that number is interpreted as a
+    /// timeinterval since 00:00:00 UTC on 1 January 1970.
+    ///
+    /// If database value contains a string, that string is interpreted as a
+    /// [SQLite date](https://sqlite.org/lang_datefunc.html) in the UTC time
+    /// zone. Nil is returned if the date string does not contain at least the
+    /// year, month and day components. Other components (minutes, etc.)
+    /// are set to zero if missing.
+    public static func fromDatabaseValue(_ dbValue: DatabaseValue) -> Date? {
+        if let databaseDateComponents = DatabaseDateComponents.fromDatabaseValue(dbValue) {
+            return Date(databaseDateComponents: databaseDateComponents)
+        }
+        if let timestamp = Double.fromDatabaseValue(dbValue) {
+            return Date(timeIntervalSince1970: timestamp)
         }
         return nil
     }
     
-    private static func fromJulianDayNumber(_ julianDayNumber: Double) -> Date? {
+    private init?(databaseDateComponents: DatabaseDateComponents) {
+        guard databaseDateComponents.format.hasYMDComponents else {
+            // Refuse to turn hours without any date information into Date:
+            return nil
+        }
+        guard let date = UTCCalendar.date(from: databaseDateComponents.dateComponents) else {
+            return nil
+        }
+        self.init(timeIntervalSinceReferenceDate: date.timeIntervalSinceReferenceDate)
+    }
+    
+    /// Creates a date from a [Julian Day](https://en.wikipedia.org/wiki/Julian_day).
+    public init?(julianDay: Double) {
         // Conversion uses the same algorithm as SQLite: https://www.sqlite.org/src/artifact/8ec787fed4929d8c
-        let JD = Int64(julianDayNumber * 86400000)
+        // TODO: check for overflows one day, and return nil when computation can't complete.
+        let JD = Int64(julianDay * 86400000)
         let Z = Int(((JD + 43200000)/86400000))
         var A = Int(((Double(Z) - 1867216.25)/36524.25))
         A = Z + 1 + A - (A/4)
@@ -52,21 +101,12 @@ extension NSDate : DatabaseValueConvertible {
         dateComponents.second = Int(second)
         dateComponents.nanosecond = Int((second - Double(Int(second))) * 1.0e9)
         
-        return UTCCalendar.date(from: dateComponents)!
-    }
-    
-    private static func fromDatabaseDateComponents(_ databaseDateComponents: DatabaseDateComponents) -> Date? {
-        guard databaseDateComponents.format.hasYMDComponents else {
-            // Refuse to turn hours without any date information into Date:
+        guard let date = UTCCalendar.date(from: dateComponents) else {
             return nil
         }
-        return UTCCalendar.date(from: databaseDateComponents.dateComponents)!
+        self.init(timeIntervalSinceReferenceDate: date.timeIntervalSinceReferenceDate)
     }
 }
-
-/// Date is stored in the database using the format
-/// "yyyy-MM-dd HH:mm:ss.SSS", in the UTC time zone.
-extension Date : DatabaseValueConvertible { }
 
 /// The DatabaseDate date formatter for stored dates.
 private let storageDateFormatter: DateFormatter = {